Implement AF_UNIX sockets on Windows
See moby/moby 36442
