+++ sockets/unix_socket_test.go
@@ -1,12 +1,9 @@
-//go:build !windows
-
 package sockets
 
 import (
 	"fmt"
 	"net"
 	"os"
-	"syscall"
 	"testing"
 )
 
@@ -52,26 +49,16 @@ func TestNewUnixSocket(t *testing.T) {
 }
 
 func TestUnixSocketWithOpts(t *testing.T) {
-	uid, gid := os.Getuid(), os.Getgid()
-	perms := os.FileMode(0o660)
-	path := "/tmp/test.sock"
-	echoStr := "hello"
-	l, err := NewUnixSocketWithOpts(path, WithChown(uid, gid), WithChmod(perms))
+	socketFile, err := os.CreateTemp("", "test*.sock")
 	if err != nil {
 		t.Fatal(err)
 	}
+	socketFile.Close()
+	defer os.Remove(socketFile.Name())
+
+	l := createTestUnixSocket(t, socketFile.Name())
 	defer l.Close()
-	p, err := os.Stat(path)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if p.Mode().Perm() != perms {
-		t.Fatalf("unexpected file permissions: expected: %#o, got: %#o", perms, p.Mode().Perm())
-	}
-	if stat, ok := p.Sys().(*syscall.Stat_t); ok {
-		if stat.Uid != uint32(uid) || stat.Gid != uint32(gid) {
-			t.Fatalf("unexpected file ownership: expected: %d:%d, got: %d:%d", uid, gid, stat.Uid, stat.Gid)
-		}
-	}
-	runTest(t, path, l, echoStr)
+
+	echoStr := "hello"
+	runTest(t, socketFile.Name(), l, echoStr)
 }
