+++ sockets/unix_socket_test_windows.go
@@ -1,0 +1,14 @@
+package sockets
+
+import (
+	"net"
+	"testing"
+)
+
+func createTestUnixSocket(t *testing.T, path string) (listener net.Listener) {
+	l, err := NewUnixSocketWithOpts(path)
+	if err != nil {
+		t.Fatal(err)
+	}
+	return l
+}
